Scruff.RTUtils

Scruff.RTUtils.RANGE — Constant

RANGE

The constant key used to store the range of a specific variable instance

Scruff.RTUtils.default_initializer — Function

Deprecated

Scruff.RTUtils.expander_range — Method

function expander_range(runtime :: Runtime, v :: Variable,
    target_size :: Int, depth :: Int)

Recursively compute the range of the expander and subnetworks up to the given depth.

Computing the range of the expander expands enough of the parent range to reach the desired target size, or expands all the parents fully.

Scruff.RTUtils.get_range — Function

get_range(runtime::Runtime, inst::Instance, depth = max_value(Int))

Returns the range value for the given instance; this will return nothing if no range has been set.

The depth specifies the maximum depth of range desired.

Scruff.RTUtils.get_range — Method

get_range(runtime::InstantRuntime, v::Node{O}, depth = 1) where O

Returns the range of the given node.

Scruff.RTUtils.get_range — Method

get_range(runtime::DynamicRuntime, v::Variable{I,J,O}, depth = 1) where {I,J,O}

Returns the range of the most recent instance of the given variable.

Scruff.RTUtils.parent_ranges — Method

parent_ranges(runtime::Runtime, var::Variable{I,J,O}, depth = typemax(Int)) where {I,J,O}

Returns the ranges of the parents of the given variable.

See Scruff.get_range

Scruff.RTUtils.set_range! — Method

set_range!(runtime::Runtime, inst::Instance{O}, range::Vector{<:O}, depth::Int = 1) where O

Sets the range value for the given instance. Defaults to depth of 1.

Scruff.RTUtils.set_ranges! — Function

set_ranges!(runtime::InstantRuntime, evidence = Dict{Symbol, Score}(),
    size = 10, depth = 1, 
    order = topsort(get_initial_graph(get_network(runtime))),
    placeholder_beliefs = get_placeholder_beliefs(runtime))

Set the ranges of all current instances in the runtime.

This method first checks whether ranges exist for the runtime at the desired depth, with the desired range size, and with the same evidence. If so, it doesn't do anything. If the depth is less than 1, it doesn't do anything. Otherwise, it uses the support operator to compute ranges of variables in order. Placeholders should have ranges set already in placeholder_beliefs. For expanders, it recursively sets the ranges of the subnetworks at depth - 1.

Returns a flag indicating whether any instance has a changed range.